fix(build): eliminate all MSVC warnings in core sources
With _CRT_SECURE_NO_WARNINGS already handling the unsafe-CRT noise, this
clears the remaining warnings so the cores compile clean under MSVC:

- C4477: print uint64_t frame counters with %llu + (unsigned long long)
  cast instead of %lu/%d (SNES bus/ppu/plugin debug logging). The old
  %lu silently truncated 64-bit counters on Windows (LLP64).
- C4244: add explicit casts at narrowing sites (GB ppu tile-addr math
  and save-state pushes, GBA ppu palette-offset accumulation, GBA bus
  HALTCNT, GBA apu SOUNDCNT_H); change the GBA RTC bcd() helper to take
  int; type the GBA flash-erase std::fill value as uint8_t.
- C4100: silence the unused is_write parameter in
  ARM7TDMI::data_access_cycles.

Also document the command-line Release build alongside Debug in the
README. Verified: both x64-Debug and x64-Release build with zero
warnings in our sources (exit 0).

+++ cores/gb/src/mbc/mbc.cpp
@@ -69,7 +69,7 @@ void MBC::write_ram(uint16_t address, uint8_t value) {
 void MBC::save_state(std::vector<uint8_t>& data) {
     data.push_back(m_rom_bank & 0xFF);
     data.push_back((m_rom_bank >> 8) & 0xFF);
-    data.push_back(m_ram_bank);
+    data.push_back(static_cast<uint8_t>(m_ram_bank));
     data.push_back(m_ram_enabled ? 1 : 0);
 }
